code/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cloudpickle==2.0.0
cycler==0.11.0
decorator==5.1.0
dm-tree==0.1.6
dpkt==1.9.8
flatbuffers==1.12
fonttools==4.28.1
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.1
Markdown==3.3.6
matplotlib==3.5.0
memory-profiler==0.61.0
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
patsy==0.5.3
Pillow==8.4.0
protobuf==3.19.1
psutil==5.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scapy==2.5.0
scikit-learn==1.1.3
scipy==1.7.2
seaborn==0.12.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
statsmodels==0.14.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-probability==0.15.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
